Organize the Gemfile 💎
We want gems to appear in alpha order, to move comments out of focus,
and to version lock production gems. The default strategy is pessimistic
lock on the patch level. Version locking the production gems in this
way allows for easier maintenence over time for an app that has few
maitainers and little developer focus.

Issue: #15