Add Groups to Organizations ⛪👨‍🏭
We need a way of describing units of service providers within a given
Org.

This change adds the `Group` class, which belongs to an `Organization`.

This changeset also enhances the test suite by configuring Factory Girl
and Shoulda Matchers.

Issue: #13

Ammended: This commit also contains files that were not added to the PR.