Add People and Association Connection 👫
The people involved in Fellowship Factory are the most important actors.
We need something to represent them.

This change introduces the `Person` class, and its associations through
`OrganizationPerson` and `GroupPerson`.

A `Person` can have many `Organization` and `Group` associations. We
have not yet enforced that the `Group` must belong to the same
`Organization` the `Person` is in.

Issue: #27